Tower 1.0

+++ Assets/CYC/Scripts/IMonster.cs
@@ -6,5 +6,5 @@ public interface IMonster
    void TakeDamage(int phyDamage, int magicDamage);
    void GetBurnt(int phyDamage, int magicDamage, int burntDamage, float burntTime);
    void GetSlow(int phyDamage, int magicDamage, float slowRatio,float slowTime);
-   void DefenseReduction(int value, float reductionTime);
+   void DefenseReduction(float value, float reductionTime);
 }
